refactor: refactor package import usage
