+++ README.md
@@ -97,7 +97,7 @@ print(resp["result"])
 
 ### 大模型调优
 
-目前千帆平台支持如下训练调优能力，详见[训练调优](./docs/console.md)
+目前千帆平台支持如下训练调优能力，详见[训练调优](./docs/train_service.md)
 - 创建训练任务
 - 创建任务运行
 - 获取任务运行详情
