+++ src/qianfan/resources/__init__.py
@@ -1,0 +1,32 @@
+from qianfan.resources.console.data import Data
+from qianfan.resources.console.finetune import FineTune
+from qianfan.resources.console.model import Model
+from qianfan.resources.console.service import Service
+from qianfan.resources.images.text2image import Text2Image
+from qianfan.resources.llm.chat_completion import ChatCompletion
+from qianfan.resources.llm.completion import Completion
+from qianfan.resources.llm.embedding import Embedding
+from qianfan.resources.llm.plugin import Plugin
+from qianfan.resources.tools.tokenizer import Tokenizer
+from qianfan.resources.typing import QfMessages, QfResponse, QfRole
+
+__all__ = [
+    "Data",
+    "Model",
+    "Service",
+    "FineTune",
+    "ChatCompletion",
+    "Embedding",
+    "Completion",
+    "Plugin",
+    "Text2Image",
+    "Tokenizer",
+    "AK",
+    "SK",
+    "Role",
+    "Messages",
+    "Response",
+    "QfRole",
+    "QfMessages",
+    "QfResponse",
+]
