+++ src/qianfan/tests/data_api_test.py
@@ -17,6 +17,7 @@
 """
 
 
+from qianfan.resources import Data
 from qianfan.resources.console.consts import (
     DataExportDestinationType,
     DataProjectType,
@@ -25,7 +26,6 @@ from qianfan.resources.console.consts import (
     DataStorageType,
     DataTemplateType,
 )
-from qianfan.resources.console.data import Data
 
 enter_except_flag = True
